+++ docsource/releasenotes.rst
@@ -24,6 +24,7 @@ UPDATED FEATURES / MAJOR BUG FIXES:
 * Removed ``kbase-admin`` credentials from the deploy.cfg file as they're obsolete after the
   conversion to auth2.
 * The credentials for the Handle Manager service in the deploy.cfg file now require a token.
+* The credentials for the file backend in the deploy.cfg file now require a token.
 
 VERSION: 0.7.0 (Released 5/5/17)
 --------------------------------